Better level-checking in constraints to handle bad bounds
Previously, we enforced level-correctness only when instantiating type
variables, but this is not good enough as demonstrated by i8900.scala (TODO:
add). The problem is that if we allow level-incorrect bounds, then we
might end up reasoning with bad bounds outside of the scope where they
are defined. This can lead to level-correct but unsound instantiations.

To prevent this, we know enforce level-correctness in constraints at all
times, we also refactored the implementation to share logic with
TypeOps#avoid (it would be worth investigating whether we could simply
drop TypeOps#avoid in favor of level-based avoidance but that would
require extra work like ensuring that TreeMaps preserve levels so I'm
leaving this for another day).

Note that this implementation is still incomplete: we only check the
nestingLevel of NamedTypes, but we also need to check for TypeVars, this
will be handled in the next commit.