Better level-checking in constraints to handle bad bounds
Previously, we enforced level-correctness only when instantiating type
variables, but this is not good enough as demonstrated by
tests/neg/i8900.scala. The problem is that if we allow level-incorrect
bounds, then we might end up reasoning with bad bounds outside of the
scope where they are defined. This can lead to level-correct but unsound
instantiations.

To prevent this, we now enforce level-correctness in constraints at all
times, we also introduce `AvoidMap` to share more logic between level-avoidance
and symbol-avoidance (see also the added TODO on `TypeOps#avoid`).

Note that this implementation is still incomplete: we only check the
nestingLevel of NamedTypes, but we also need to check for TypeVars, this
will be handled in the next commit.