Fix OrderingConstraint#order forgetting some constraints
`order` takes `current` as input and adds `param1 <: param2` to it, it should be
idempotent: if `param1 <: param2` is already true, there is nothing to be done
and we can return `current` as is. But the current implementation instead
incorrectly returned `this`, which could very well be a different, older
constraint! In practice this almost never happens