Experiment: forced instantiation only uses non-param bound
Type variable are normally either instantiated to their lower bound (after some
widening in `widenInferred`) or upper bound. This commit changes this behavior
when the type variable instantiation is forced (which happens in particular
before doing an implicit search, but also when typing the parameters of a
lambda). We now use the "nonParam" bound instead where other type parameters
have been stripped. For example in i16328.scala, when typing

    id(fromParam((x: Base) => x))

we end up with

    id[?S](getParamType[?T]((x: Int) => x))

where
    ?T <: ?S
    ?T <: Base

Before this commit, ?T was then instantiated to ?S & Base (via
Typer#adaptNoArgsImplicitMethod#instantiate), which lead
to an implicit search failure since ?S was left unconstrained.

With this commit, we instead instantiate ?T to its non-param bound `Base`
and propagate this to `?S` itself:

   ?T := Base
   ?S >: Base

The idea is that this leads to constraints being propagated in a more intuitive
way (?S will now be instantiated to Base rather than Nothing or Any), however
this also means that we're adding a whole new way of instantiating type
variables for a seemingly limited number of usecases (see also the discussion
starting at
https://github.com/lampepfl/dotty/pull/16328#issuecomment-1317401760).

It's not clear if this won't end up breaking as many things as it fixes, and it
would increase our maintenance burden, so I recommend holding off on this change
for now.