Value parameter inference for polymorphic closures
Just like
    val f: Int => Int = x => x
infers a parameter type of Int, we can now infer the binder x to have type T in:
    val g: [T] => T => T = [T] => x => x

Since the `T` in the expected type and the closure are different, this requires
a substitution step which is tricky to do since we're operating with untyped
trees at this point. We implement this by generalizing the existing
`DependentTypeTree` mechanism (already used for computing dependent result types
of lambdas) to also be usable in any other position inside the lambda.

We rename the tree to `InLambdaTypeTree` at the same time for clarity.

Note that this mechanism could also probably be used to allow closures with
internal value parameter dependencies, but we don't attempt to support this
here.